+++ AzzyBot/Modules/MusicStreaming/MusicStreamingModule.cs
@@ -8,14 +8,10 @@ internal sealed class MusicStreamingModule : BaseModule
 {
     internal override void RegisterCommands(SlashCommandsExtension slashCommandsExtension, ulong? serverId) => slashCommandsExtension.RegisterCommands<MusicStreamingCommands>(serverId);
 
-    protected override async void HandleModuleEvent(ModuleEvent evt)
+    protected override void HandleModuleEvent(ModuleEvent evt)
     {
         switch (evt.Type)
         {
-            case ModuleEventType.LavalinkPassword:
-                evt.ResultString = await MusicStreamingLavalinkHandler.GetLavalinkPasswordAsync();
-                break;
-
             case ModuleEventType.GetMusicStreamingInactivity:
                 evt.ResultBool = MusicStreamingSettings.AutoDisconnect;
                 break;
@@ -32,6 +28,9 @@ internal sealed class MusicStreamingModule : BaseModule
 
     internal override async void StartProcesses()
     {
+        if (GetAzzyBotName() is "AzzyBot-Docker")
+            return;
+
         if (!await MusicStreamingLavalinkHandler.CheckIfJavaIsInstalledAsync())
             throw new InvalidOperationException("You have to install Java/OpenJDK Runtime 17 or 21 first!");
 
@@ -41,4 +40,11 @@ internal sealed class MusicStreamingModule : BaseModule
 
     internal override async void StopProcesses() => await MusicStreamingLavalinkHandler.StopLavalinkAsync();
     internal override void Activate() => ModuleStates.ActivateMusicStreaming();
+
+    internal static string GetAzzyBotName()
+    {
+        ModuleEvent evt = new(ModuleEventType.GetAzzyBotName);
+        BroadcastModuleEvent(evt);
+        return evt.ResultString;
+    }
 }